Pull customized alttp generator into a separate include
